requirements.txt
Django==1.11.5
djangorestframework==3.8.2
mysqlclient==1.3.10
PyMySQL==0.9.2
requests==2.19.1
django-filter==1.1.0
djangorestframework-filters==0.10.2.post0
mysql-replication==0.18

